cpukit/libcsupport/src/pwdgrp.c: Address -Wsign-compare warnings
This warning occurs when comparing a signed variable to an unsigned one.
This addresses warnings that only occurred on 64-bit targets.  For the
ones which only appeared on 64-bit targets, the cause was frequently
a mismatch when comparing a combination off_t, ssize_t, and int.

+++ cpukit/libcsupport/src/pwdgrp.c
@@ -163,7 +163,7 @@ static int scanInt( FILE *fp, int *val )
     d = c - '0';
     if (
       ( i > ( limit / 10 ) ) ||
-      ( ( i == ( limit / 10 ) ) && ( d > ( limit % 10 ) ) )
+      ( ( i == ( limit / 10 ) ) && ( d > (unsigned int)( limit % 10 ) ) )
     ) {
       return 0;
     }
